fixed已经被half淘汰了

+++ shader-urp/Assets/code/shaders/04.tessellate.shader
@@ -36,11 +36,11 @@ Shader "core/04.tessellate"
 			float2 uv_Albedo;
 		};
 
-		fixed4		_AlbedoColor;
+		half4		_AlbedoColor;
 		sampler2D	_Albedo;
 		half		_Bumpiness;
 		sampler2D 	_Normal;
-		fixed4 		_EmissionColor;
+		half4 		_EmissionColor;
 		sampler2D 	_Emission;
 
 		half _Glossiness;
@@ -69,7 +69,7 @@ Shader "core/04.tessellate"
 		void surf(Input input , inout SurfaceOutput output)
 		{
             // ONE
-            fixed4 c = tex2D(_Albedo, input.uv_Albedo) * _AlbedoColor; // 讲道理每一个texture都有一个乘参
+            half4 c = tex2D(_Albedo, input.uv_Albedo) * _AlbedoColor; // 讲道理每一个texture都有一个乘参
             output.Albedo = c.rgb;	//  albedo与alpha是一对, 加起来恰好是一个float4, 因此albedo是rgb
 			output.Alpha = c.a;
